% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_munich_forecast_query.R
\name{weather_munich_forecast_query}
\alias{weather_munich_forecast_query}
\title{Title}
\usage{
weather_munich_forecast_query(forecast_days = 14, past_days = 0)
}
\arguments{
\item{past_days}{}
}
\description{
Title
}
\keyword{internal}
