% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_munich_get.R
\name{weather_munich_get}
\alias{weather_munich_get}
\title{Get Munich weather forecast data as tibble}
\usage{
weather_munich_get(forecast_days = 7, past_days = 0)
}
\arguments{
\item{forecast_days}{A number between 0 and 16}

\item{past_days}{A number between 0 and 82}
}
\value{
A tibble
}
\description{
Queries the Open-Meteo.com API for Munich and returns daily weather data for selected number of days in the past and in the future as a tibble.
\if{html}{\out{<a href="https://open-meteo.com/">}}Weather data by Open-Meteo.com\if{html}{\out{</a>}}
}
\examples{
weather_munich_get(forecast_days = 7, past_days = 0)
}
