% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_munich_plot_temp3m.R
\name{weather_munich_plot_temp3m}
\alias{weather_munich_plot_temp3m}
\title{Temperature Weather Plot}
\usage{
weather_munich_plot_temp3m(forecast_days = 14, past_days = 0)
}
\arguments{
\item{forecast_days}{A number between 0 and 16}

\item{past_days}{A number between 0 and 82}
}
\value{
A ggplot2 plot object
}
\description{
Function that returns a ggplot2 plot visualizing the daily minimum, maximum and
mean temperatures for the requested days. Can be modified further using "+" in the usual
ggplot2 fashion, if needed.
}
\examples{
weather_munich_plot_temp3m(forecast_days = 14, past_days = 0)
}
